evals/pyproject.toml
# Simple script-based project for dependency management only
[project]
name = "policy-analyzer-evals"
version = "0.1.0"
description = "Golden Dataset Evaluation System for Policy Analyzer"
authors = [
    {name = "Policy Analyzer Team"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.1.0",
    "langchain-experimental>=0.0.50",
    "openai>=1.0.0",
    "ragas>=0.1.0",
    "qdrant-client>=1.7.0",
    "datasets>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyMuPDF>=1.23.0",
    "Pillow>=9.0.0",
    "langsmith>=0.1.0",
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 